requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
setuptools==74.1.2
sqlparse==0.5.1